Seperate external links editor into own component; improve page UI
