+++ singadrive-web-client/components/userprofile/editor/EditorContent.tsx
@@ -17,11 +17,15 @@ const DescriptionEditor = dynamic(() => import('@/components/userprofile/editor/
   ssr: false
 });
 
+const ExternalLinksEditor = dynamic(() => import('@/components/userprofile/editor/ExternalLinksEditor'), {
+  ssr: false
+});
+
 const ProfileEditor: React.FC = () => {
   const [username, setUsername] = useState<string>('');
   const [displayName, setDisplayName] = useState('');
   const [birthday, setBirthday] = useState<Date | null>(null);
-  const [externalLinks, setExternalLinks] = useState<ExternalLink[]>([]);
+  const [externalLinks, setExternalLinks] = useState<{ title: string; url: string }[]>([]);
   const [email, setEmail] = useState('');
   const [newPassword, setNewPassword] = useState('');
   const [description, setNewDescription] = useState<string>('');
@@ -83,86 +87,72 @@ const ProfileEditor: React.FC = () => {
   }, []);
 
   return (
-    <div className="profile-editor">
-      <h1>Edit Profile</h1>
+      <>
+        <h1>Edit Profile</h1>
+        <div className="profile-editor">
 
-      <div className="input-group">
-        <label>Username</label>
-        <span>{username}</span>
-      </div>
+          <div className="input-group">
+            <div className="input-label">
+              <label>Username</label>
+            </div>
+            <div className="input-control">
+              <span>{username}</span>
+            </div>
+          </div>
 
-      <div className="input-group">
-        <label htmlFor="displayName">Display Name</label>
-        <input
-          type="text"
-          id="displayName"
-          value={displayName}
-          onChange={(e) => setDisplayName(e.target.value)}
-        />
-      </div>
-      
-      <div className="input-group">
-        <label>Birthday</label>
-        <DatePicker selected={birthday} onChange={(date: Date) => setBirthday(date)} />
-      </div>
-      
-      <div className="input-group">
-        <label>External Links</label>
-        {externalLinks.map((link, index) => (
-          <div key={index}>
+          <div className="input-group">
+            <label htmlFor="displayName">Display Name</label>
             <input
               type="text"
-              placeholder="Title"
-              value={link.title}
-              onChange={(e) => {
-                const newLinks = [...externalLinks];
-                newLinks[index].title = e.target.value;
-                setExternalLinks(newLinks);
-              }}
-            />
-            <input
-              type="url"
-              placeholder="URL"
-              value={link.url}
-              onChange={(e) => {
-                const newLinks = [...externalLinks];
-                newLinks[index].url = e.target.value;
-                setExternalLinks(newLinks);
-              }}
+              id="displayName"
+              value={displayName}
+              onChange={(e) => setDisplayName(e.target.value)}
             />
           </div>
-        ))}
-        <button onClick={() => setExternalLinks([...externalLinks, { title: '', url: '' }])}>Add Link</button>
+
+          <div className="input-group">
+            <div className="input-label">
+              <label>Birthday</label>
+            </div>
+          </div>
+          <div className="input-group">
+            <div className="input-control">
+              <DatePicker selected={birthday} onChange={(date: Date) => setBirthday(date)} />
+            </div>
+          </div>
+
+
+          <ExternalLinksEditor externalLinks={externalLinks} setExternalLinks={setExternalLinks} /> 
+
+          <div className="input-group">
+            <label>Profile Description</label>
+          </div>
+          <DescriptionEditor initialDescription='' onDescriptionChange={setNewDescription}/>
+
+          <hr/>
+          <div className="input-group">
+            <label htmlFor="email">Email</label>
+            <input
+              type="email"
+              id="email"
+              value={email}
+              onChange={(e) => setEmail(e.target.value)}
+            />
+          </div>
+
+          <div className="input-group">
+            <label htmlFor="newPassword">New Password</label>
+            <input
+              type="password"
+              id="newPassword"
+              value={newPassword}
+              onChange={(e) => setNewPassword(e.target.value)}
+            />
+          </div>
+          <hr/>
+          <button type="submit">Save Changes</button>
       </div>
-
-        <div className="input-group">
-          <label htmlFor="email">Email</label>
-          <input
-            type="email"
-            id="email"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-          />
-        </div>
-
-        <div className="input-group">
-          <label htmlFor="newPassword">New Password</label>
-          <input
-            type="password"
-            id="newPassword"
-            value={newPassword}
-            onChange={(e) => setNewPassword(e.target.value)}
-          />
-        </div>
-
-        <div className="input-group">
-          <label>Description</label>
-        </div>
-        <DescriptionEditor initialDescription='' onDescriptionChange={setNewDescription}/>
-
-        <hr/>
-        <button type="submit">Save Changes</button>
-    </div>
+    </>
   );
 };
